osi/mem: fix transmute test on big-endian
Ensure that our const-test for transmute_copy_uninit() does not rely on
endianness. Just use a no-op transmute instead.

+++ lib/osi/src/mem.rs
@@ -218,7 +218,7 @@ mod test {
     // Verify `transmute_copy_uninit()` works in constant contexts.
     #[test]
     fn transmute_copy_uninit_const() {
-        const C: u16 = unsafe { transmute_copy_uninit(&71u32) };
+        const C: u32 = unsafe { transmute_copy_uninit(&71u32) };
 
         assert_eq!(C, 71);
     }
